Attoparsec port
This implements ghc-events on top of attoparsec instead of binary. The
advantage is that we gain a lot of freedom in how to handle errors
(see getEvents).

Also note readEventLogFromFile: Now we can parse data in chunks, and we
could theoretically get progressive results by feeding the partial parse
result EOFs (empty bytestrings), which would cause most events parsed
so far to be returned.

Finally, note that when parsing block markers, we can not simply
recursively call getEvents, as attoparsec has no means to reason about the
amount of input we're consuming - so there's no way to tell when we have
reached the end of a block.